# platform/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/alerts": {
      "get": {
        "summary": "List Alerts",
        "operationId": "list_alerts_api_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "type": "array",
                  "title": "Response List Alerts Api Alerts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Alert",
        "operationId": "create_alert_api_alerts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{alert_id}": {
      "put": {
        "summary": "Update Alert",
        "operationId": "update_alert_api_alerts__alert_id__put",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Alert",
        "operationId": "delete_alert_api_alerts__alert_id__delete",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/coin/{instrument_id}/indicators": {
      "get": {
        "summary": "Get Coin Indicator Config",
        "description": "Return this instrument's persisted picker selection. No saved entry is a legitimate\n\"nothing saved yet\" state -- `[]`, not an error.",
        "operationId": "get_coin_indicator_config_api_coin__instrument_id__indicators_get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IndicatorConfigEntry"
                  },
                  "title": "Response Get Coin Indicator Config Api Coin  Instrument Id  Indicators Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Coin Indicator Config",
        "description": "Persist this instrument's current indicator selection (explicit Save/change, never\nauto-save-per-keystroke -- matches today's handler's docstring). The body is a JSON array of\n`IndicatorConfigEntry` (declared through `openapi_extra`).\n\nReads the raw JSON body itself (rather than a typed Pydantic body parameter) so a\nmalformed entry -- missing `name`/`category`, or invalid JSON outright -- degrades to a\nreal `400` with a message, never FastAPI's automatic `422`, matching the pre-existing\nhandler's contract exactly (DATA-02: never a silent/opaque failure for a client-input\nproblem). An unknown name or params the indicator refuses are 400 too, a source the\nindicator cannot take 422; nothing is written then. A corrupt or unreadable stored file is\n500.",
        "operationId": "put_coin_indicator_config_api_coin__instrument_id__indicators_put",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Put Coin Indicator Config Api Coin  Instrument Id  Indicators Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IndicatorConfigEntry"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertCreate": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "once_per_bar_close",
              "once_per_bar",
              "only_once"
            ],
            "title": "Frequency"
          },
          "bar_seconds": {
            "type": "integer",
            "title": "Bar Seconds"
          },
          "expires_at_ns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Ns"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "webhook_url": {
            "type": "string",
            "title": "Webhook Url"
          },
          "condition": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/LevelCondition"
                  },
                  {
                    "$ref": "#/components/schemas/PctMoveCondition"
                  },
                  {
                    "$ref": "#/components/schemas/ChannelExitCondition"
                  },
                  {
                    "$ref": "#/components/schemas/IndicatorCondition"
                  },
                  {
                    "$ref": "#/components/schemas/TrendlineCrossCondition"
                  },
                  {
                    "$ref": "#/components/schemas/FundingCondition"
                  },
                  {
                    "$ref": "#/components/schemas/OiChangeCondition"
                  },
                  {
                    "$ref": "#/components/schemas/LiquidationNotionalCondition"
                  },
                  {
                    "$ref": "#/components/schemas/ForcedShareCondition"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "channel_exit": "#/components/schemas/ChannelExitCondition",
                    "forced_share": "#/components/schemas/ForcedShareCondition",
                    "funding_above": "#/components/schemas/FundingCondition",
                    "funding_below": "#/components/schemas/FundingCondition",
                    "indicator": "#/components/schemas/IndicatorCondition",
                    "liquidation_notional": "#/components/schemas/LiquidationNotionalCondition",
                    "oi_change": "#/components/schemas/OiChangeCondition",
                    "pct_move": "#/components/schemas/PctMoveCondition",
                    "price_above": "#/components/schemas/LevelCondition",
                    "price_below": "#/components/schemas/LevelCondition",
                    "price_cross": "#/components/schemas/LevelCondition",
                    "price_cross_down": "#/components/schemas/LevelCondition",
                    "price_cross_up": "#/components/schemas/LevelCondition",
                    "trendline_cross": "#/components/schemas/TrendlineCrossCondition"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "frequency",
          "bar_seconds",
          "template",
          "webhook_url"
        ],
        "title": "AlertCreate"
      },
      "AlertResponse": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "once_per_bar_close",
              "once_per_bar",
              "only_once"
            ],
            "title": "Frequency"
          },
          "bar_seconds": {
            "type": "integer",
            "title": "Bar Seconds"
          },
          "expires_at_ns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Ns"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "webhook_url": {
            "type": "string",
            "title": "Webhook Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "triggered",
              "expired",
              "invalid"
            ],
            "title": "Status"
          },
          "created_ns": {
            "type": "integer",
            "title": "Created Ns"
          },
          "last_fired_ns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fired Ns"
          },
          "condition": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LevelCondition"
              },
              {
                "$ref": "#/components/schemas/PctMoveCondition"
              },
              {
                "$ref": "#/components/schemas/ChannelExitCondition"
              },
              {
                "$ref": "#/components/schemas/IndicatorCondition"
              },
              {
                "$ref": "#/components/schemas/TrendlineCrossCondition"
              },
              {
                "$ref": "#/components/schemas/FundingCondition"
              },
              {
                "$ref": "#/components/schemas/OiChangeCondition"
              },
              {
                "$ref": "#/components/schemas/LiquidationNotionalCondition"
              },
              {
                "$ref": "#/components/schemas/ForcedShareCondition"
              }
            ],
            "title": "Condition",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "channel_exit": "#/components/schemas/ChannelExitCondition",
                "forced_share": "#/components/schemas/ForcedShareCondition",
                "funding_above": "#/components/schemas/FundingCondition",
                "funding_below": "#/components/schemas/FundingCondition",
                "indicator": "#/components/schemas/IndicatorCondition",
                "liquidation_notional": "#/components/schemas/LiquidationNotionalCondition",
                "oi_change": "#/components/schemas/OiChangeCondition",
                "pct_move": "#/components/schemas/PctMoveCondition",
                "price_above": "#/components/schemas/LevelCondition",
                "price_below": "#/components/schemas/LevelCondition",
                "price_cross": "#/components/schemas/LevelCondition",
                "price_cross_down": "#/components/schemas/LevelCondition",
                "price_cross_up": "#/components/schemas/LevelCondition",
                "trendline_cross": "#/components/schemas/TrendlineCrossCondition"
              }
            }
          },
          "condition_text": {
            "type": "string",
            "title": "Condition Text"
          },
          "invalid_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalid Reason"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "frequency",
          "bar_seconds",
          "template",
          "webhook_url",
          "id",
          "status",
          "created_ns",
          "condition",
          "condition_text"
        ],
        "title": "AlertResponse"
      },
      "AlertUpdate": {
        "properties": {
          "condition": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LevelCondition"
              },
              {
                "$ref": "#/components/schemas/PctMoveCondition"
              },
              {
                "$ref": "#/components/schemas/ChannelExitCondition"
              },
              {
                "$ref": "#/components/schemas/IndicatorCondition"
              },
              {
                "$ref": "#/components/schemas/TrendlineCrossCondition"
              },
              {
                "$ref": "#/components/schemas/FundingCondition"
              },
              {
                "$ref": "#/components/schemas/OiChangeCondition"
              },
              {
                "$ref": "#/components/schemas/LiquidationNotionalCondition"
              },
              {
                "$ref": "#/components/schemas/ForcedShareCondition"
              }
            ],
            "title": "Condition",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "channel_exit": "#/components/schemas/ChannelExitCondition",
                "forced_share": "#/components/schemas/ForcedShareCondition",
                "funding_above": "#/components/schemas/FundingCondition",
                "funding_below": "#/components/schemas/FundingCondition",
                "indicator": "#/components/schemas/IndicatorCondition",
                "liquidation_notional": "#/components/schemas/LiquidationNotionalCondition",
                "oi_change": "#/components/schemas/OiChangeCondition",
                "pct_move": "#/components/schemas/PctMoveCondition",
                "price_above": "#/components/schemas/LevelCondition",
                "price_below": "#/components/schemas/LevelCondition",
                "price_cross": "#/components/schemas/LevelCondition",
                "price_cross_down": "#/components/schemas/LevelCondition",
                "price_cross_up": "#/components/schemas/LevelCondition",
                "trendline_cross": "#/components/schemas/TrendlineCrossCondition"
              }
            }
          },
          "frequency": {
            "type": "string",
            "enum": [
              "once_per_bar_close",
              "once_per_bar",
              "only_once"
            ],
            "title": "Frequency"
          },
          "expires_at_ns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Ns"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "webhook_url": {
            "type": "string",
            "title": "Webhook Url"
          },
          "rearm": {
            "type": "boolean",
            "title": "Rearm",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "condition",
          "frequency",
          "template",
          "webhook_url"
        ],
        "title": "AlertUpdate",
        "description": "`PUT /api/alerts/{id}`: the editable fields (the instrument and bar width are the alert's\nidentity and stay). `rearm` clears a triggered `only_once` alert's `triggered`."
      },
...
      "ChannelExitCondition": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "channel_exit",
            "title": "Kind"
          },
          "upper": {
            "type": "number",
            "title": "Upper"
          },
          "lower": {
            "type": "number",
            "title": "Lower"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "upper",
          "lower"
        ],
        "title": "ChannelExitCondition"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndicatorConfigEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "close"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden",
            "default": false
          },
          "style": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Style",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "title": "IndicatorConfigEntry"
      },
...
      "LevelCondition": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "price_cross",
              "price_cross_up",
              "price_cross_down",
              "price_above",
              "price_below"
            ],
            "title": "Kind"
          },
          "level": {
            "type": "number",
            "title": "Level"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "level"
        ],
        "title": "LevelCondition"
      },
...
      "PctMoveCondition": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "pct_move",
            "title": "Kind"
          },
          "pct": {
            "type": "number",
            "title": "Pct"
          },
          "bars": {
            "type": "integer",
            "title": "Bars"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "pct",
          "bars"
        ],
        "title": "PctMoveCondition"
      },
...
    }
  }
}